packages/workspaces/default_sync_backup/pyproject.toml
[project]
name = "ai-stock-research"
version = "0.1.0"
description = "AI Stock Research Institute - Multi-agent stock analysis platform"
requires-python = ">=3.11"
dependencies = [
    # Web framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    # Database
    "sqlalchemy[asyncio]>=2.0.30",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    # Cache
    "redis[hiredis]>=5.0.0",
    # Message queue
    "aiokafka>=0.10.0",
    # Configuration
    "pydantic-settings>=2.3.0",
    # HTTP client
    "httpx>=0.27.0",
    # Logging
    "structlog>=24.1.0",
    # Resilience & Rate Limiting
    "tenacity>=8.4.0",
    "aiolimiter>=1.1.0",
    # Scheduler
    "apscheduler>=3.10.4",
    "holidays>=0.49",
    "beautifulsoup4>=4.14.3",
    "lxml>=6.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["services", "libs", "configs"]

[dependency-groups]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "pre-commit>=3.7.0",
]

# ── Testing ──────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "unit: Unit tests (no external dependencies)",
    "integration: Integration tests (require infrastructure)",
]

# ── Linting & Formatting ────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length = 100
src = ["services", "libs", "configs", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["services", "libs", "configs"]

# ── Type Checking ────────────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
mypy_path = "."
packages = ["services", "libs", "configs"]
